change leveldbclusteringdata using Table and add the reload function

+++ source/core/laser-manager/service/RpcServer.cpp
@@ -66,8 +66,8 @@ bool RpcServer::init(const std::string& clusteringRootPath,
 }
 
 void RpcServer::start(const std::string& host, 
-        uint16_t port, 
-        uint32_t threadNum, 
+        int port, 
+        int threadNum, 
         const std::string& clusteringRootPath, 
         const std::string& clusteringDBRootPath, 
         const std::string& pcaPath)
